Simplification of RaceConfigurerSetOne and added a Registry
Registry contains the details of all leagues, teams etc. and provides a
means for passing a comparotor in order to determine seeds. This is by
design i.e. the registry only knows about what entites exist and how
they are ranked, scored etc. This allows scoring, and therefore seeding,
to be kept independent.

This mechanism has been used to significantly reduce the logic in
RaceConfigurerSetOne so that only requires the result of calling this
method on the registry for input.

It should be noted though that this is only a start and there are other
layers which are still of significant complexity, not to mention the
other RaceConfigurer.